fixing testing bugs
