+++ tests/__main__.py
@@ -6,3 +6,5 @@ def main():
     """
     print("To run all tests, use `python3 -m unittest discover tests` in the home directory!")
     return 0
+
+main()
