+++ tests/test_polltypes.py
@@ -8,7 +8,6 @@ import random
 sys.path.append(os.path.dirname(__file__)) #gets pdoc working
 
 from shared.pollTypes import *
-from shared.database import *
 
 
 
@@ -150,133 +149,7 @@ class PollMethodsTesting(unittest.TestCase):
     def test_toBytes(self):
         self.assertEqual(self.myPoll.toBytes(), json.dumps(self.d).encode())
 
-class DatabaseSQLTesting(unittest.TestCase):
 
-    @classmethod
-    def setUpClass(cls):
-
-        #remove test db
-        try:
-            os.remove("./shared/testingDB.db")
-        except FileNotFoundError:
-            pass
-        #load up db
-        cls.db = DatabaseSQL("./shared/testingDB.db")
-
-        cls.userTemplate = {
-        "test@gmail.com" : {
-                "firstName" : "John",
-                "lastName" : "Doe",
-                "password" : "ecd4d1aad41a446759d25de6c830d60cc3c8548be9760f0babe03094e6a59ee3",
-                "classes" : [12, 1442, 123],
-                "reedemed" : True,
-                "role" : "professors"
-            }
-        }
-
-        cls.classTemplate = {
-            "className": 'Into to Blah',
-            "courseCode": 'UNIQ99',
-            "students" : ["mrstudent@gmail.com"],
-            "professors" : ["mrprof@gmail.com"],
-            "polls": []
-        }
-
-        cls.pollTemplate = {
-            'question': {
-                'prompt': 'What is your favorite color?',
-                "answer" : None, 
-                'options': [], 
-                'type': 'FreeResponseQuestion'
-                },
-            'startTime' : "2020-10-27 16:25:07",
-            'endTime' : "2025-10-27 16:25:07",
-            'ownerId' : 'bebop@yahoo.com',
-            'classId' : 0,
-            'responses': []}
-        
-
-    @classmethod
-    def tearDownClass(cls):
-        pass
-
-    def test_addStudentToDB(self):
-        student = self.userTemplate
-
-        #test add user to db
-        self.assertTrue(self.db.addUser(student))
-
-        #test pull that user from the db
-        result = self.db.getUser("test@gmail.com")
-        self.assertEqual(student, result)
-
-        #test pull to nothing
-        result = self.db.getUser("notindb")
-        self.assertFalse(result)
-
-        #test change value
-        result = self.db.updateFieldViaId("users", "test@gmail.com", "firstName", "NewName")
-        self.assertTrue(result)
-        student["test@gmail.com"]["firstName"] = "NewName"
-
-        self.assertEqual(student, self.db.getUser("test@gmail.com"))
-    
-    def test_addClassToDB(self):
-        classD = self.classTemplate
-
-        #add class to db
-        self.assertTrue(self.db.addClass(classD))
-
-        #Pull class via id
-        result = self.db.getClassFromId(1)
-        self.assertEqual(classD, result)
-
-        #pull class via course id
-        result = self.db.getClassFromCourseCode("UNIQ99")
-        self.assertEqual(classD, result[1])
-
-        #test change value
-        result = self.db.updateFieldViaId("classes", 1, "courseCode", "YEE555")
-        self.assertTrue(result)
-        classD["courseCode"] = "YEE555"
-
-        self.assertEqual(classD, self.db.getClassFromId(1))
-
-    def test_addPollToDB(self):
-        d = self.pollTemplate
-
-        poll = Poll.fromDict(d)
-        self.assertTrue(self.db.addPoll(poll))
-
-        self.assertEqual(d, self.db.getPollFromId(1))
-
-
-    def test_addResponseToDB(self):
-        pass
-
-    def test_getUsersOfClass(self):
-        """This test asserts that profs and students can be successfully indexed from classes"""
-        self.maxDiff = None
-        newStudent = self.userTemplate
-        newStudent["mrstudent@gmail.com"] = newStudent.pop(list(newStudent.keys())[0])
-        newStudent[list(newStudent.keys())[0]]['role'] = "students"
-        self.db.addUser(newStudent)
-
-        newProf = self.userTemplate
-        newProf["mrprof@gmail.com"] = newProf.pop(list(newProf.keys())[0])
-        newProf[list(newStudent.keys())[0]]['firstName'] = "ProfJohn"
-        newProf[list(newStudent.keys())[0]]['role'] = "professors"
-        self.db.addUser(newProf)
-
-
-        st = self.db.getStudentsInClass(1)
-        self.assertEqual(len(st.keys()), 1)
-        self.assertEqual(list(st.keys())[0], "mrstudent@gmail.com")
-
-
-        pf = self.db.getProfsInClass(1)
-        self.assertEqual(len(pf.keys()), 1)
-        self.assertEqual(list(pf.keys())[0], "mrprof@gmail.com")
 
 
 if __name__ == '__main__':
